requirements.txt
urllib3 == 1.15
environs == 4.2.0
python-json-logger == 0.1.11
Flask == 0.10
pytest

# Common
loguru ~= 0.5.3
requests ~= 2.25.1
pytz ~= 2021.1
pycryptodome == 3.10.1
pyopenssl == 19.1.0
pymysql ~= 1.0.2
sqlalchemy ~= 1.4.17
sqlalchemy-pagination ~= 0.0.2
paginator == 0.5.1
inflection == 0.5.1
arrow == 1.1.0
sqlalchemy-utils == 0.37.5
jsonschema~=3.2.0

# GRPC libs`
# six~=1.10
# wheel~=0.29
# protobuf
# grpcio-tools~=1.38.0
# grpcio==1.38.0
# coverage~=4.0
# cython~=0.29.8
